--- src/transformers/models/minicpm_o_2_6/tokenization_minicpm_o_2_6.py
@@ -1,24 +1,0 @@
-# coding=utf-8
-# Copyright 2025 The OpenBMB Team. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from transformers import Qwen2Tokenizer
-
-
-class MiniCPM_o_2_6Tokenizer(Qwen2Tokenizer):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-
-__all__ = ["MiniCPM_o_2_6Tokenizer"]
